validate category children on delete

+++ sources/backend/categories/Services/Delete/Service.cs
@@ -19,6 +19,11 @@ namespace Elesse.Categories
          if (category == null)
             return Warning(WARNINGS.CATEGORY_NOT_FOUND);
 
+         // VALIDATE IF CATEGORY HAS NO CHILDREN
+         var children = await _CategoryRepository.SearchCategoriesAsync(category.Type, category.CategoryID, "");
+         if (children?.Length > 0)
+            return Warning(WARNINGS.STILL_HAS_CHILDREN_CANT_REMOVE);
+
          // REMOVE CATEGORY
          await _CategoryRepository.DeleteCategoryAsync(categoryID);
 
@@ -39,6 +44,7 @@ namespace Elesse.Categories
    partial struct WARNINGS
    {
       internal const string INVALID_DELETE_PARAMETER = "INVALID_CATEGORYID_PARAMETER";
+      internal const string STILL_HAS_CHILDREN_CANT_REMOVE = "STILL_HAS_CHILDREN_CANT_REMOVE";
    }
 
 }
